requirements.txt
tornado
sqlalchemy
numpy
configparser
py-sc-client==0.2.0
